--- src/addcommentToDom.js
@@ -1,8 +1,0 @@
-const addCommentToDom = (comments) => {
-  const commentList = document.querySelector('.comments-container');
-  commentList.innerHTML = '';
-  comments.forEach((comment) => {
-    commentList.innerHTML += `<li>${comment.creation_date}: ${comment.username}: ${comment.comment}</li>`;
-  });
-};
-export default addCommentToDom;
